+++ Client/src/containers/avs-video-converter-other-content.js
@@ -1,5 +1,6 @@
 import React from "react";
 import Link from '../components/link';
+import { Trans } from "react-i18next";
 import {Link as ScrollLink} from "react-scroll";
 import Text from '../components/text';
 import FreeFlag from '../components/free-flag';
@@ -97,6 +98,28 @@ export const AvsVideoConverterOtherContent = (props) => (
         <PageContentWrapper className="videoConverter">
             <ContentRowItem
                 id="overview"
+                imgLeft={false}
+                asType={props.t("h2")}
+                newText={props.t("New")}
+                imageName={props.t("video converter convert video with ai png")}
+                headerText={props.t("Convert video with AI")}
+                showFirstBlockBullets={false}
+                secondHeaderText={props.t("Add new AI–generated effects")}
+                showSecondBlockBullets={false}
+                secondBlockContent={
+                    <Text className="ListItem">
+                        <Trans
+                            i18nKey="Add AI-powered image and video effects while preparing your video"
+                            components={[
+                                <a key="0" target="_blank" href="https://avs4you.github.io" />,
+                            ]}
+                        />
+                    </Text>
+                }
+            >
+                <Text className="ListItem">{props.t("Go from a text idea to a perfectly formatted video")}</Text>
+            </ContentRowItem>
+            <ContentRowItem
                 imgLeft={true}
                 asType={props.t("h2")}
                 imageName={props.t("video converter convert all key2 jpg")}
@@ -273,7 +296,7 @@ export const AvsVideoConverterOtherContent = (props) => (
             </div>
         </div>
 
-        <Text id="screenshotsCarousel" className="carouselHeader" as="h2">AVS Video Converter 14.1</Text>
+        <Text id="screenshotsCarousel" className="carouselHeader" as="h2">AVS Video Converter 26.0</Text>
 
 
         <ScreenshotsCarousel